refactor(llma): move playground logic into separate files
Split the monolithic llmAnalyticsPlaygroundLogic into three helper
files (prompts, model, run) under a new playground/ directory. The
main llmPlaygroundLogic re-exports and composes them via spread.
